move `key` package to internal
